+++ 03_infra/ts_parser.rs
@@ -168,6 +168,7 @@ impl<R: PromptReader, S: PromptSnapshotReader> LanguageParser for TsParser<R, S>
             language: file.language.clone(),
             prompt_header,
             prompt_file_exists,
+            prompt_refs: vec![],
             has_test_coverage,
             imports,
             tokens,